[MOB-656] Remove empty initializers for Customer and PaymentMethod
The empty initializers allow instantiation of a Customer and PaymentMethod without any data at all. Since all the properties were optional, this is possible. However, this also means that it's possible to parse _anything_ into a Customer. So network requests that return an error were successfully being parsed as instances of Customer and PaymentMethod, which caused legitimate errors to be silently thrown away