Make MobileReader.connectionType public
ChipDnaDriver assumes a connectionType of 'Bluetooth', however when trying to connect to a BLE device using Bluetooth, the connection doesn't work.
When connectionType is internal, it doesn't let someone assign a value to it, which means that someone can't create a MobileReader instance from scratch.
This is a big deal for apps built with hybrid mobile app frameworks since they are often having to serialize our objects to pass them between langs